+++ src/app/core/entity-components/entity-details/form/form.component.ts
@@ -14,6 +14,7 @@ import {
 } from "../../entity-form/entity-form.service";
 import { AlertService } from "../../../alerts/alert.service";
 import { toFormFieldConfig } from "../../entity-subrecord/entity-subrecord/entity-subrecord-config";
+import * as _ from "lodash-es";
 
 /**
  * A simple wrapper function of the EntityFormComponent which can be used as a dynamic component
@@ -25,12 +26,14 @@ import { toFormFieldConfig } from "../../entity-subrecord/entity-subrecord/entit
   templateUrl: "./form.component.html",
   styleUrls: ["./form.component.scss"],
 })
-export class FormComponent implements OnInitDynamicComponent, OnInit {
-  entity: Entity;
+export class FormComponent<E extends Entity>
+  implements OnInitDynamicComponent, OnInit
+{
+  entity: E;
   columns: FormFieldConfig[][] = [];
   headers?: string[] = [];
   creatingNew = false;
-  form: EntityForm<Entity>;
+  form: EntityForm<E>;
 
   constructor(
     private router: Router,
@@ -40,7 +43,7 @@ export class FormComponent implements OnInitDynamicComponent, OnInit {
   ) {}
 
   onInitFromDynamicConfig(config: PanelConfig) {
-    this.entity = config.entity;
+    this.entity = config.entity as E;
     this.columns = config.config?.cols.map((row) => row.map(toFormFieldConfig));
     this.headers = config.config?.headers;
     if (config.creatingNew) {
@@ -64,7 +67,10 @@ export class FormComponent implements OnInitDynamicComponent, OnInit {
       this.form.markAsPristine();
       this.form.disable();
       if (this.creatingNew) {
-        this.router.navigate([getParentUrl(this.router), this.entity.getId()]);
+        await this.router.navigate([
+          getParentUrl(this.router),
+          this.entity.getId(),
+        ]);
       }
     } catch (err) {
       if (!(err instanceof InvalidFormFieldError)) {
@@ -84,6 +90,10 @@ export class FormComponent implements OnInitDynamicComponent, OnInit {
   private resetForm(entity = this.entity) {
     // Patch form with values from the entity
     this.form.patchValue(entity as any);
+    const newKeys = Object.keys(
+      _.omit(this.form.controls, Object.keys(this.entity))
+    );
+    newKeys.forEach((key) => this.form.get(key).setValue(null));
     this.form.markAsPristine();
   }
 }
